requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
elevenlabs==2.9.1
openai==1.3.7
pydantic==2.5.0